Rename gateway packages and use upstream AG-UI SDK
