+++ apps/messenger-webhook/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@agui/messenger-webhook",
+  "name": "@agui-gw/fb-messenger",
   "version": "0.1.0",
   "private": true,
   "description": "Facebook Messenger webhook gateway for AG-UI",
@@ -13,6 +13,7 @@
     "ioredis": "^5.3.2",
     "pino": "^9.3.2",
     "prom-client": "^15.1.2",
-    "zod": "^3.22.4"
+    "zod": "^3.22.4",
+    "@agui-gw/core": "workspace:*"
   }
 }
