+++ packages/gateway-core/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@agui-gw/core",
+  "version": "0.1.0",
+  "description": "Core building blocks shared by Messenger gateway services",
+  "type": "module",
+  "license": "MIT",
+  "sideEffects": false,
+  "exports": {
+    ".": {
+      "import": "./src/index.ts",
+      "types": "./src/index.ts"
+    }
+  },
+  "dependencies": {
+    "@ag-ui/client": "^0.1.0",
+    "@ag-ui/core": "^0.1.0",
+    "@ag-ui/encoder": "^0.1.0",
+    "@ag-ui/proto": "^0.1.0"
+  }
+}
