+++ packages/gateway-core/src/index.ts
@@ -1,0 +1,8 @@
+export type {
+  AguiDispatcher,
+  DispatchContext,
+  AguiDispatchHandlers,
+  AguiDispatcherOptions,
+} from './agui/dispatcher';
+export { createAguiDispatcher, buildRunInput } from './agui/dispatcher';
+export type { AssistantMessagePayload, RunLifecyclePayload, RunErrorPayload } from '@ag-ui/core';
