+++ packages/gateway-core/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist",
+    "noEmit": true,
+    "composite": false
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": ["dist"]
+}
